LayoutTests/platform/ios-simulator/ios/scrolling/scrollTo-at-page-load-expected.txt

Tests that programmatic scrolling works at page load time.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received scroll event 1
PASS scrollEventCount is 1
PASS window.scrollY is 10
PASS successfullyParsed is true

TEST COMPLETE
